style(lib/fetchgeocoords): rename file to it's exported class name
By convention, a file that exports only one class, function, or constant should be named for that
class, function or constant. Anything else may confuse maintainers.
